Migrate Query.myDefaultOrganization
